AddEduApp/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
django-dramatiq==0.11.6
django-environ==0.11.2
django-pwa==2.0.0
django-webpush==0.3.6
dramatiq==1.17.0
et-xmlfile==1.1.0
frozenlist==1.4.1
http-ece==1.2.0
idna==3.7
multidict==6.0.5
openpyxl==3.1.5
pika==1.3.2
pillow==10.3.0
prometheus_client==0.20.0
psycopg2==2.9.9
py-vapid==1.9.1
pycparser==2.22
PyJWT==2.8.0
pywebpush==2.0.0
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.6.0
yarl==1.9.4
